backend/pyproject.toml
[project]
name = "indeksygsr"
version = "0.1.0"
description = "Wyszukiwarka semantyczna indeksów materiałowych GSR"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client==2.190.0",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.2.4",
    "libpff-python-ratom==20220304",
    "pandas==2.3.3",
    "google-cloud-firestore==2.23.0",
    "google-cloud-storage==3.9.0",
    "streamlit==1.54.0",
    "playwright==1.58.0",
    "FlagEmbedding==1.3.5",
    "transformers>=4.40,<5.0",
    "qdrant-client==1.17.0",
    "python-dotenv==1.2.1",
    "spacy==3.8.11",
    "pl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_lg-3.8.0/pl_core_news_lg-3.8.0-py3-none-any.whl",
]

[project.scripts]
scrape      = "app.pipeline.scrape:main"
parse-email = "app.pipeline.parse_email:main"
vectorize   = "app.pipeline.vectorize:main"
search      = "app.core.search:main"
gdrive      = "app.services.gdrive:main"
dashboard   = "app.dashboard:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
